tests: reconciliation: reorganize
